+++ src/OptionsResolver/PaymentMethodsSettingsResolver.php
@@ -33,6 +33,7 @@ class PaymentMethodsSettingsResolver extends AbstractSettingsResolver
             ->setDefined([
                 'displayGenericOption',
                 'genericLogoFilename',
+                'isDefaultGenericLogo',
                 'deleteGenericLogo',
                 'redirectTemplateFilename',
                 'redirectCallToAction',
@@ -44,6 +45,7 @@ class PaymentMethodsSettingsResolver extends AbstractSettingsResolver
                 'iframePaymentMethods',
                 'iframeLogo',
                 'iframeLogoFilename',
+                'isDefaultIframeLogo',
                 'enabled',
                 'productId',
                 'identifier',