+++ translations/es.php
@@ -136,7 +136,7 @@ $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_bafd7322c6e97d25b6299
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_002c18d9b9b0bcb5eb2888cdfeb6373a'] = 'Un solo botón de pago que redirecciona a una página alojada de Worldline para realizar el pago';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ffaaab40fc54c578836d4ae92238718c'] = 'Logotipo genérico que se muestra en su página de pago';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_91412465ea9169dfd901dd5e7c96dd99'] = 'Cargar';
-$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_8de30aa607e33f41413c50be107e0c6c'] = 'Eliminar el logotipo actual';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_0a7aa3e6a43a0a1406e4169ae2be963a'] = 'Restaurar logotipo predeterminado';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_254ad5cf512e99709dbcb70646ed7b6d'] = 'Aquí puedes cargar un logotipo nuevo (los únicos tipos de archivo admitidos para los logotipos son .png, .gif y .jpg)';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ff76e4c4c9abdc7de325c97912ffb8cd'] = 'Recomendamos utilizar imágenes con un tamaño máximo de 20 píxeles de altura por 120 píxeles de anchura';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_3c8990c77ed9e1634b5cdd133c6e0993'] = 'Título del botón de pago';