+++ translations/fr.php
@@ -136,7 +136,7 @@ $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_bafd7322c6e97d25b6299
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_002c18d9b9b0bcb5eb2888cdfeb6373a'] = 'Un bouton de paiement unique pour être redirigé pour payer sur une seule page hébergée par Worldline';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ffaaab40fc54c578836d4ae92238718c'] = 'Logo générique affiché sur votre page de paiement';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_91412465ea9169dfd901dd5e7c96dd99'] = 'Télécharger';
-$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_8de30aa607e33f41413c50be107e0c6c'] = 'Supprimer le logo actuel';
+$_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_0a7aa3e6a43a0a1406e4169ae2be963a'] = 'Restaurer le logo par défaut';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_254ad5cf512e99709dbcb70646ed7b6d'] = 'Vous pouvez télécharger un nouveau logo ici (les types de fichier acceptés pour les logos sont uniquement : .png .gif .jpg)';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_ff76e4c4c9abdc7de325c97912ffb8cd'] = 'Nous vous recommandons d’utiliser des images ayant une hauteur maximale de 20 px et une longueur maximale de 120 px';
 $_MODULE['<{worldlineop}prestashop>_paymentmethodssettings_3c8990c77ed9e1634b5cdd133c6e0993'] = 'Titre du bouton de paiement';