+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.16] - 2025-14-11
+### Fixed
+- Fix being able to delete default logo when upgrading from earlier version
+
+## [1.4.15] - 2025-13-11
+### Added
+- Add reverting to default logo in payment methods settings page
+
 ## [1.4.14] - 2025-30-10
 ### Changed
 - Fix transaction operations having wrong status